[feature/DB-125]: 리뷰 API 수정 (#35)
* fix: 검색 조건 변경
* refactor(DB-126): 리뷰 수정
* refactor: 리뷰 내용 보기 type sort로 변환
* refactor: 소나 수정
* fix: 북마크 조회 정렬 변경
* fix: BookmarkQueryDSLRepository getBookmarks test 수정
* fix: 자신이 작성한 산책로 검색 허용, 리뷰 보기 허용

+++ infrastructure/storage-rdb/src/main/java/com/dongsan/domains/review/dto/RatingCount.java
@@ -1,7 +1,7 @@
 package com.dongsan.domains.review.dto;
 
 public record RatingCount(
-        Byte rating,
+        Integer rating,
         Long count
 ) {
 }
